infra/local-dev/services/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.2
PyJWT[crypto]==2.8.0
cryptography==42.0.5
